+++ packages/hooks/src/index.ts
@@ -15,3 +15,6 @@ export { useHotkeys } from './useHotkeys';
 export { useDidUpdate } from './useDidUpdate';
 export { useMergedRef, mergeRefs } from './useMergedRef';
 export { useDebouncedValue } from './useDebouncedValue';
+export { useDidMount } from './useDidMount';
+export { useDidUpdate } from './useDidUpdate';
+export { useUnmount } from './useUnmount';